requirements.txt
absl-py @ file:///opt/conda/conda-bld/absl-py_1639803114343/work
aiohttp @ file:///C:/ci/aiohttp_1602530358780/work
astor==0.8.1
astunparse==1.6.3
async-timeout==3.0.1
attrs @ file:///opt/conda/conda-bld/attrs_1642510447205/work
blinker==1.4
brotlipy==0.7.0
cachetools @ file:///tmp/build/80754af9/cachetools_1619597386817/work
certifi==2022.6.15
cffi @ file:///C:/ci_310/cffi_1642682485096/work
chardet @ file:///C:/ci/chardet_1605303259695/work
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
click @ file:///C:/ci/click_1646038601470/work
colorama @ file:///C:/Windows/TEMP/abs_9439aeb1-0254-449a-96f7-33ab5eb17fc8apleb4yn/croots/recipe/colorama_1657009099097/work
cryptography @ file:///C:/ci/cryptography_1652101770956/work
dnspython==1.16.0
email-validator @ file:///tmp/build/80754af9/email-validator_1623674793554/work
Flask @ file:///C:/ci/flask_1658860110736/work
Flask-SQLAlchemy @ file:///home/conda/feedstock_root/build_artifacts/flask-sqlalchemy_1616102249273/work
Flask-WTF @ file:///tmp/build/80754af9/flask-wtf_1622018667443/work
gast==0.3.3
google-auth @ file:///opt/conda/conda-bld/google-auth_1646735974934/work
google-auth-oauthlib @ file:///tmp/build/80754af9/google-auth-oauthlib_1617120569401/work
google-pasta @ file:///Users/ktietz/demo/mc3/conda-bld/google-pasta_1630577991354/work
greenlet @ file:///D:/bld/greenlet_1648882598553/work
grpcio @ file:///C:/ci/grpcio_1637590993074/work
gunicorn==20.1.0
h5py==2.10.0
idna @ file:///tmp/build/80754af9/idna_1637925883363/work
importlib-metadata @ file:///C:/ci/importlib-metadata_1648562631189/work
itsdangerous @ file:///tmp/build/80754af9/itsdangerous_1621432558163/work
Jinja2 @ file:///opt/conda/conda-bld/jinja2_1647436528585/work
Keras-Applications @ file:///tmp/build/80754af9/keras-applications_1594366238411/work
Keras-Preprocessing @ file:///tmp/build/80754af9/keras-preprocessing_1612283640596/work
Markdown @ file:///C:/ci/markdown_1614364005059/work
MarkupSafe @ file:///C:/ci/markupsafe_1654508076077/work
mkl-fft==1.3.1
mkl-random @ file:///C:/ci/mkl_random_1626186163140/work
mkl-service==2.4.0
multidict @ file:///C:/ci/multidict_1600456486794/work
numpy @ file:///C:/ci/numpy_and_numpy_base_1653574840943/work
oauthlib==3.1.0
opt-einsum @ file:///tmp/build/80754af9/opt_einsum_1621500238896/work
Pillow==9.0.1
protobuf==3.19.1
psycopg2==2.9.3
pyasn1 @ file:///Users/ktietz/demo/mc3/conda-bld/pyasn1_1629708007385/work
pyasn1-modules==0.2.8
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
PyJWT==1.7.1
pyOpenSSL @ file:///opt/conda/conda-bld/pyopenssl_1643788558760/work
pyreadline==2.1
PySocks @ file:///C:/ci/pysocks_1594394709107/work
requests @ file:///opt/conda/conda-bld/requests_1641824580448/work
requests-oauthlib==1.3.0
rsa @ file:///tmp/build/80754af9/rsa_1614366226499/work
scipy @ file:///C:/ci/scipy_1641555141383/work
six @ file:///tmp/build/80754af9/six_1644875935023/work
SQLAlchemy @ file:///D:/bld/sqlalchemy_1659998874450/work
tensorboard @ file:///home/builder/ktietz/aggregate/tensorflow_recipes/ci_te/tensorboard_1614593728657/work/tmp_pip_dir
tensorboard-plugin-wit==1.6.0
tensorflow==2.3.0
tensorflow-estimator @ file:///home/builder/adipietro/tf/tensorflow-estimator_1630508970172/work/tensorflow_estimator-2.6.0-py2.py3-none-any.whl
termcolor==1.1.0
typing_extensions @ file:///C:/Windows/TEMP/abs_dd2d0moa85/croots/recipe/typing_extensions_1659638831135/work
urllib3 @ file:///C:/ci/urllib3_1650639842077/work
Werkzeug @ file:///opt/conda/conda-bld/werkzeug_1645628268370/work
win-inet-pton @ file:///C:/ci/win_inet_pton_1605306165655/work
wincertstore==0.2
wrapt @ file:///C:/ci/wrapt_1638415942355/work
WTForms @ file:///Users/ktietz/demo/mc3/conda-bld/wtforms_1629354749392/work
yapf @ file:///tmp/build/80754af9/yapf_1615749224965/work
yarl @ file:///C:/ci/yarl_1606940114600/work
zipp @ file:///C:/ci/zipp_1652274072582/work
